requirements.txt
fastapi
uvicorn
python-multipart
python-jose
passlib[bcrypt]

langchain-core
langchain-community
langchain-text-splitters
faiss-cpu
sentence-transformers
transformers
torch
pypdf
python-docx

# --- OCR & PDF Processing ---
pdf2image
paddlepaddle>=2.6.2
paddleocr>=2.7.0
opencv-python-headless

# ... các thư viện cũ ...
llama-index-core
llama-index-embeddings-huggingface
llama-index-vector-stores-faiss
llama-index-readers-file

fastapi
uvicorn
pydantic
pydantic-settings
filelock
psutil
faiss-cpu
pytest
